v1.8.0 for Minecraft 1.9.4
Update to Minecraft 1.9.4 (still need to fix double command bug for
right click only)
