--- plugin.yml
@@ -1,5 +1,0 @@
-name: CommandNPC
-author: messageofdeath
-version: 1.8.0
-main: me.razorrider7.commandnpc.CommandNPC
-depend: [Citizens]
