+++ src/main/java/me/razorrider7/commandnpc/database/ClickType.java
@@ -1,0 +1,10 @@
+package me.razorrider7.commandnpc.database;
+
+public 	enum ClickType {
+	
+	Left(),
+	
+	Right(),
+	
+	Both()
+}
