+++ src/main/resources/config.yml
@@ -1,0 +1,1 @@
+ClickType: Interact
