--- src/me/messageofdeath/CommandNPC/Database/ClickType.java
@@ -1,10 +1,0 @@
-package me.messageofdeath.CommandNPC.Database;
-
-public 	enum ClickType {
-	
-	Left(),
-	
-	Right(),
-	
-	Both()
-}
